+++ tests/ui/range.stderr
@@ -1,42 +1,10 @@
-error: Iterator::step_by(0) will panic at runtime
-  --> $DIR/range.rs:8:13
-   |
-LL |     let _ = (0..1).step_by(0);
-   |             ^^^^^^^^^^^^^^^^^
-   |
-   = note: `-D clippy::iterator-step-by-zero` implied by `-D warnings`
-
-error: Iterator::step_by(0) will panic at runtime
-  --> $DIR/range.rs:12:13
-   |
-LL |     let _ = (1..).step_by(0);
-   |             ^^^^^^^^^^^^^^^^
-
-error: Iterator::step_by(0) will panic at runtime
-  --> $DIR/range.rs:13:13
-   |
-LL |     let _ = (1..=2).step_by(0);
-   |             ^^^^^^^^^^^^^^^^^^
-
-error: Iterator::step_by(0) will panic at runtime
-  --> $DIR/range.rs:16:13
-   |
-LL |     let _ = x.step_by(0);
-   |             ^^^^^^^^^^^^
-
 error: It is more idiomatic to use v1.iter().enumerate()
-  --> $DIR/range.rs:24:14
+  --> $DIR/range.rs:5:14
    |
 LL |     let _x = v1.iter().zip(0..v1.len());
    |              ^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = note: `-D clippy::range-zip-with-len` implied by `-D warnings`
 
-error: Iterator::step_by(0) will panic at runtime
-  --> $DIR/range.rs:28:13
-   |
-LL |     let _ = v1.iter().step_by(2 / 3);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^
-
-error: aborting due to 6 previous errors
+error: aborting due to previous error
